refactor(data-model): move JSON-encoding area into src/json-wire/
Reorganize the JSON-encoding classes and non-class bits into a new
`src/json-wire/` directory (sibling of `fabric-instances/` and
`fabric-primitives/`), one file per class plus one file per handler
singleton, behind a single barrel subpath export.

- `json-type-handlers.ts` decomposed into: `TypeHandlerRegistry.ts`
  (class); the 8 handler singletons (`UndefinedHandler.ts`,
  `BigIntHandler.ts`, `EpochNsecHandler.ts`, `EpochDaysHandler.ts`,
  `SpecialNumberHandler.ts`, `SymbolHandler.ts`, `BytesHandler.ts`,
  `FabricInstanceHandler.ts`), each in a file whose base matches its
  export name; `json-wire-types.ts` (shared `JsonWireValue` /
  `TypeHandler` / `TypeHandlerCodec` vocabulary); `make-problematic.ts`
  (the formerly file-private `makeProblematic`, now an exported shared
  helper); and `create-default-registry.ts` (the `createDefaultRegistry`
  factory).
- `JsonEncodingContext.ts` and `json-encoding.ts` moved into
  `json-wire/` (near-pure moves; module-private helpers travel with
  them). Bodies are byte-identical; only import blocks change.
- `json-wire/index.ts` barrel is the sole public entry. `deno.json`
  replaces the three per-file specifiers with a single `./json-wire`.
- Cross-package importers (memory, runner test, toolshed x2,
  runtime-client) repointed to `@commonfabric/data-model/json-wire`;
  the two intra-package test files repointed to the moved `src/json-wire`
  paths.
