+++ packages/data-model/deno.json
@@ -38,9 +38,7 @@
     "./value-clone": "./src/value-clone.ts",
     "./frozen-builtins": "./src/frozen-builtins.ts",
     "./interface": "./src/interface.ts",
-    "./json-encoding": "./src/json-encoding.ts",
-    "./JsonEncodingContext": "./src/JsonEncodingContext.ts",
-    "./json-type-handlers": "./src/json-type-handlers.ts",
+    "./json-wire": "./src/json-wire/index.ts",
     "./native-type-tags": "./src/native-type-tags.ts",
     "./schema-hash": "./src/schema-hash.ts",
     "./schema-utils": "./src/schema-utils.ts",
