+++ packages/data-model/src/json-wire/FabricInstanceHandler.ts
@@ -1,0 +1,62 @@
+import {
+  DECONSTRUCT,
+  FabricInstance,
+  type FabricValue,
+  type ReconstructionContext,
+} from "../interface.ts";
+import { ExplicitTagValue } from "../fabric-instances/ExplicitTagValue.ts";
+import type {
+  JsonWireValue,
+  TypeHandler,
+  TypeHandlerCodec,
+} from "./json-wire-types.ts";
+
+/**
+ * Handler for `FabricInstance` values (custom protocol types, including
+ * `FabricError` and `ExplicitTagValue` subtypes). Serializes via
+ * `[DECONSTRUCT]` and the codec's tag methods. Deserialization is not
+ * dispatched via this handler's tag (since each instance type has its own
+ * tag like `TAGS.Error`); instead, the deserializer falls back to the class
+ * registry for those tags.
+ */
+export const FabricInstanceHandler: TypeHandler = {
+  // This tag is not used for deserialization dispatch -- `FabricInstance`
+  // types are looked up by their individual tags. The handler is registered
+  // for serialization matching only.
+  tag: "",
+
+  canSerialize(value: FabricValue): boolean {
+    return value instanceof FabricInstance;
+  },
+
+  serialize(
+    value: FabricValue,
+    codec: TypeHandlerCodec,
+    recurse: (v: FabricValue) => JsonWireValue,
+  ): JsonWireValue {
+    const inst = value as FabricInstance;
+
+    // `ExplicitTagValue` (`UnknownValue`, `ProblematicValue`): use
+    // preserved `.typeTag` and re-serialize their stored state.
+    if (inst instanceof ExplicitTagValue) {
+      const serializedState = recurse(inst.state);
+      return codec.wrapTag(inst.typeTag, serializedState);
+    }
+
+    // General `FabricInstance`: use `[DECONSTRUCT]` and codec for tag.
+    const state = inst[DECONSTRUCT]();
+    const tag = codec.getTagFor(inst);
+    const serializedState = recurse(state);
+    return codec.wrapTag(tag, serializedState);
+  },
+
+  deserialize(
+    _state: JsonWireValue,
+    _runtime: ReconstructionContext,
+    _recurse: (v: JsonWireValue) => FabricValue,
+  ): FabricValue {
+    // Not reached via tag dispatch -- `FabricInstance` deserialization is
+    // handled by the class registry fallback in `deserialize()`.
+    throw new Error("FabricInstanceHandler.deserialize should not be called");
+  },
+};
