+++ packages/runtime-client/backends/runtime-processor.ts
@@ -1,7 +1,7 @@
 import { DID, Identity, type Session } from "@commonfabric/identity";
 import { FabricBytes } from "@commonfabric/data-model/fabric-primitives";
 import type { FabricValue } from "@commonfabric/data-model/fabric-value";
-import { JsonEncodingContext } from "@commonfabric/data-model/JsonEncodingContext";
+import { JsonEncodingContext } from "@commonfabric/data-model/json-wire";
 import { PieceManager } from "@commonfabric/piece";
 import { PiecesController } from "@commonfabric/piece/ops";
 import {
