+++ packages/toolshed/routes/blobs/blobs.index.ts
@@ -7,7 +7,7 @@ import { memoryServer } from "@/routes/storage/memory.ts";
 import { isDID } from "@commonfabric/identity";
 import { FabricBytes } from "@commonfabric/data-model/fabric-primitives";
 import { hashOf } from "@commonfabric/data-model/value-hash";
-import { JsonEncodingContext } from "@commonfabric/data-model/JsonEncodingContext";
+import { JsonEncodingContext } from "@commonfabric/data-model/json-wire";
 import { EmptyReconstructionContext } from "@commonfabric/data-model/EmptyReconstructionContext";
 import {
   decodeMemoryBoundary,
